feat: Final class containers

+++ containers/container.h
@@ -7,7 +7,7 @@ public:
  virtual void push_back(T value) = 0;
  virtual void insert(int pos, T value) = 0;
  virtual void erase(int pos) = 0;
- virtual int size() = 0;
+ virtual size_t size() = 0;
  virtual T operator[](int index) = 0;
  virtual void print() = 0;
 };
